Feat: Change game vitrine to show images in a carousel
