+++ app/src/main/java/com/github/meeplemeet/ui/shops/ShopScreen.kt
@@ -44,6 +44,8 @@ import com.github.meeplemeet.model.shops.ShopViewModel
 import com.github.meeplemeet.ui.components.AvailabilitySectionWithChevron
 import com.github.meeplemeet.ui.components.ContactSection
 import com.github.meeplemeet.ui.components.GameDetailsCard
+import com.github.meeplemeet.ui.components.GameImageListSection
+import com.github.meeplemeet.ui.components.GameItemImage
 import com.github.meeplemeet.ui.components.ImageCarousel
 import com.github.meeplemeet.ui.components.ShopComponentsTestTags
 import com.github.meeplemeet.ui.components.TopBarWithDivider
@@ -70,7 +72,7 @@ object ShopScreenDefaults {
     const val GAME_SECTION_TITLE = "Discover Games"
     val GAME_NAME_AREA_HEIGHT = 46.dp
     const val GAME_NAME_MAX_LINES = 2
-    const val GAME_IMG_RELATIVE_WIDTH = 0.85f
+    const val GAME_IMG_RELATIVE_WIDTH = 0.75f
     const val GAME_IMG_DEFAULT_ASPECT_RATIO = 0.75f
   }
 
@@ -239,196 +241,3 @@ fun ShopDetails(
       }
 }
 
-// -------------------- GAME ITEM --------------------
-
-/**
- * A composable function that displays a game item as an image card with an optional stock badge
- *
- * @param game The [Game] object whose image and name are displayed
- * @param count The stock quantity for the game. When greater than zero, a stock badge is shown
- * @param modifier The [Modifier] to be applied to the root container of the game item
- * @param clickable A boolean indicating whether the game card is clickable
- * @param onClick A callback function that is invoked when the game card is clicked
- * @param imageHeight An optional fixed height for the image area
- */
-@Composable
-fun GameItemImage(
-    game: Game,
-    count: Int,
-    modifier: Modifier = Modifier,
-    clickable: Boolean = true,
-    onClick: (Game) -> Unit = {},
-    imageHeight: Dp? = null,
-) {
-
-  Box(
-      modifier =
-          modifier
-              .fillMaxWidth()
-              .testTag("${ShopComponentsTestTags.SHOP_GAME_PREFIX}${game.uid}")) {
-        Column(
-            modifier =
-                Modifier.fillMaxWidth()
-                    .padding(top = ShopScreenDefaults.Stock.STOCK_BUBBLE_TOP_PADDING)
-                    .clip(MaterialTheme.shapes.medium)
-                    .background(MaterialTheme.colorScheme.background)
-                    .clickable(enabled = clickable) { onClick(game) }
-                    .padding(Dimensions.Padding.small),
-            horizontalAlignment = Alignment.CenterHorizontally) {
-              Box(
-                  modifier =
-                      Modifier.fillMaxWidth(ShopScreenDefaults.Game.GAME_IMG_RELATIVE_WIDTH)
-                          .shadow(
-                              Dimensions.Elevation.high, MaterialTheme.shapes.medium, clip = true)
-                          .clip(MaterialTheme.shapes.medium)
-                          .background(MaterialTheme.colorScheme.background)
-                          .let { base ->
-                            if (imageHeight != null) base.height(imageHeight)
-                            else
-                                base.aspectRatio(
-                                    ShopScreenDefaults.Game.GAME_IMG_DEFAULT_ASPECT_RATIO)
-                          }) {
-                    AsyncImage(
-                        model = game.imageURL,
-                        contentDescription = game.name,
-                        contentScale = ContentScale.Crop,
-                        modifier = Modifier.fillMaxSize(),
-                        placeholder = painterResource(R.drawable.ic_dice),
-                        error = painterResource(R.drawable.ic_dice))
-                  }
-
-              Spacer(Modifier.height(Dimensions.Spacing.small))
-
-              Text(
-                  text = game.name,
-                  style = MaterialTheme.typography.bodySmall,
-                  textAlign = TextAlign.Center,
-                  maxLines = ShopScreenDefaults.Game.GAME_NAME_MAX_LINES,
-                  overflow = TextOverflow.Ellipsis,
-                  modifier =
-                      Modifier.fillMaxWidth()
-                          .testTag("${ShopTestTags.SHOP_GAME_NAME_PREFIX}${game.uid}"))
-            }
-
-        if (count > ShopScreenDefaults.Stock.NOT_SHOWING_STOCK_MIN_VALUE) {
-          val label =
-              if (count > ShopScreenDefaults.Stock.MAX_STOCK_SHOWED)
-                  "${ShopScreenDefaults.Stock.MAX_STOCK_SHOWED}+"
-              else count.toString()
-
-          Box(
-              modifier =
-                  Modifier.size(ShopScreenDefaults.Stock.STOCK_BUBBLE_SIZE)
-                      .align(Alignment.TopStart)
-                      .offset(x = Dimensions.Padding.small, y = Dimensions.Padding.small)
-                      .clip(CircleShape)
-                      .background(MaterialTheme.colorScheme.primary)
-                      .testTag("${ShopTestTags.SHOP_GAME_STOCK_PREFIX}${game.uid}"),
-              contentAlignment = Alignment.Center) {
-                Text(
-                    text = label,
-                    style = MaterialTheme.typography.bodySmall,
-                    color = MaterialTheme.colorScheme.onPrimary)
-              }
-        }
-      }
-}
-
-/**
- * A composable function that displays a paged grid section of game image items with an optional
- * title and page indicators
- *
- * @param games The list of pairs of [Game] and stock count to display in the grid
- * @param modifier The [Modifier] to be applied to the section container
- * @param clickableGames A boolean indicating whether individual game cards are clickable
- * @param title The title text displayed above the grid (for example, "Discover Games")
- * @param onClick A callback function that is invoked when a game card is clicked
- */
-@SuppressLint("UnusedBoxWithConstraintsScope")
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun GameImageListSection(
-    games: List<Pair<Game, Int>>,
-    modifier: Modifier = Modifier,
-    clickableGames: Boolean = false,
-    title: String,
-    onClick: (Game) -> Unit = {},
-) {
-  val clampedGames = remember(games) { games.shuffled().take(ShopScreenDefaults.Pager.MAX_GAMES) }
-  if (clampedGames.isEmpty()) return
-
-  val pages =
-      remember(clampedGames) {
-        clampedGames
-            .chunked(ShopScreenDefaults.Pager.GAMES_PER_PAGE)
-            .take(ShopScreenDefaults.Pager.MAX_PAGES)
-      }
-  val pageCount = pages.size
-
-  val pagerState = rememberPagerState(pageCount = { pageCount })
-  val scope = rememberCoroutineScope()
-
-  Column(
-      verticalArrangement = Arrangement.spacedBy(Dimensions.Spacing.extraSmall),
-      modifier = modifier.fillMaxWidth()) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleLarge,
-            fontWeight = FontWeight.SemiBold,
-        )
-
-        BoxWithConstraints(modifier = Modifier.fillMaxWidth()) {
-          val gridWidth = maxWidth
-          val imageHeight = gridWidth / ShopScreenDefaults.Pager.IMAGE_HEIGHT_CORRECTION
-          val textAreaHeight = ShopScreenDefaults.Game.GAME_NAME_AREA_HEIGHT
-          val rowHeight = imageHeight + textAreaHeight
-          val gridHeight = rowHeight * ShopScreenDefaults.Pager.GAMES_PER_COLUMN
-
-          HorizontalPager(
-              state = pagerState,
-              modifier =
-                  Modifier.fillMaxWidth()
-                      .height(gridHeight)
-                      .testTag(ShopTestTags.SHOP_GAME_PAGER)) { pageIndex ->
-                LazyVerticalGrid(
-                    columns = GridCells.Fixed(ShopScreenDefaults.Pager.GAMES_PER_ROW),
-                    horizontalArrangement = Arrangement.spacedBy(Dimensions.Spacing.medium),
-                    userScrollEnabled = false,
-                    modifier = Modifier.fillMaxSize()) {
-                      items(pages[pageIndex], key = { it.first.uid }) { (game, count) ->
-                        GameItemImage(
-                            game = game,
-                            count = count,
-                            clickable = clickableGames,
-                            onClick = onClick,
-                            imageHeight = imageHeight,
-                            modifier = Modifier.height(rowHeight),
-                        )
-                      }
-                    }
-              }
-        }
-
-        if (pageCount > ShopScreenDefaults.Pager.MINIMAL_PAGE_COUNT) {
-          Row(
-              modifier = Modifier.fillMaxWidth().padding(top = Dimensions.Spacing.medium),
-              horizontalArrangement = Arrangement.Center) {
-                repeat(pageCount) { index ->
-                  val selected = (index == pagerState.currentPage)
-                  Box(
-                      modifier =
-                          Modifier.padding(horizontal = Dimensions.Padding.small)
-                              .size(
-                                  if (selected) ShopScreenDefaults.Pager.PAGER_SELECTED_BUBBLE_SIZE
-                                  else ShopScreenDefaults.Pager.PAGER_UNSELECTED_BUBBLE_SIZE)
-                              .clip(CircleShape)
-                              .testTag("${ShopTestTags.SHOP_GAME_PAGER_INDICATOR_PREFIX}$index")
-                              .background(
-                                  if (selected) MaterialTheme.colorScheme.primary
-                                  else MaterialTheme.colorScheme.outline)
-                              .clickable { scope.launch { pagerState.animateScrollToPage(index) } })
-                }
-              }
-        }
-      }
-}
